+++ src/helpers/handleHTML.js
@@ -1,7 +1,14 @@
-import * as DOMPurify from 'dompurify'
+import DOMPurify from 'dompurify'
 import parse from 'html-react-parser'
 
 export const removeHTMLTags = (text) => text.replace(/(<([^>]+)>)/gi, '')
+
+DOMPurify.addHook('afterSanitizeAttributes', function (node) {
+    if (node.tagName.toLowerCase() === 'a') {
+        node.setAttribute('target', '_blank')
+        node.setAttribute('rel', 'noreferrer')
+    }
+})
 
 export const sanitizeMainHTML = (html) =>
     DOMPurify.sanitize(html, {