+++ src/pages/login/index.js
@@ -1,0 +1,3 @@
+export { InnerLoginForm } from './InnerLoginForm.jsx'
+export { LoginErrors } from './LoginErrors.jsx'
+export { LoginForm } from './LoginForm.jsx'